% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-partitions.R
\name{check_slurm_partitions}
\alias{check_slurm_partitions}
\title{throws error if the number of requested CPUs exceeds
the number of CPUs available in the requested partition}
\usage{
check_slurm_partitions(ncpu, partition, cache = TRUE)
}
\arguments{
\item{ncpu}{number of CPUs requested by user}

\item{partition}{name of partition requested by user}

\item{cache}{optional argument to forgo caching}
}
\description{
throws error if the number of requested CPUs exceeds
the number of CPUs available in the requested partition
}
\examples{
check_slurm_partitions(17, "cpu2mem4gb")
check_slurm_partitions(3, "cpu2mem4gb")
check_slurm_partitions(5, "cpu4mem32gb")
check_slurm_partitions(5, "cpu4mem32gb")
check_slurm_partitions(100, "cpu32mem128gb")
check_slurm_partitions(2, "cpu2mem4gb")
}
