% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmm-config.R
\name{generate_nmm_config}
\alias{generate_nmm_config}
\title{Creates a default watcher configuration file}
\usage{
generate_nmm_config(
  .mod,
  model_number = "",
  files_to_track = c("lst", "ext", "grd"),
  tmp_dir = "/tmp",
  watched_dir = file.path("model", "nonmem"),
  output_dir = file.path(watched_dir, "in_progress"),
  poll_duration = 1,
  alert = "None",
  level = "Debug",
  email = "",
  threads = 1
)
}
\arguments{
\item{.mod}{a bbi nonmem model object}

\item{model_number}{a string of model number e.g. 101a15. Default is pulled from .mod}

\item{files_to_track}{a vector of file extensions. Default is c("lst", "ext", "grd")}

\item{tmp_dir}{a temporary directory location to run nonmem. Default is /tmp}

\item{watched_dir}{the directory where nonmem will output. Default is here::here()/model/nonmem}

\item{output_dir}{the directory where watcher will place log and any output files default is here::here()/model/nonmem/in_progress}

\item{poll_duration}{the amount of time in seconds between the watchers polling}

\item{alert}{Where to send alerts if any. Default is None, available options are None or Slack}

\item{level}{What level to log at. Default is info. Available options are Trace, Debug, Info, Warn, Fatal}

\item{email}{if alert is set to Slack, this should be the email associated with slack to get messages sent directly to you.}

\item{threads}{number of threads if running a parallel job. Default is 1}
}
\value{
none
}
\description{
Creates a default watcher configuration file
}
\examples{
\dontrun{
  generate_nmm_config(.mod)
}
}
\keyword{internal}
