% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_watcher_config.R
\name{generate_watcher_config}
\alias{generate_watcher_config}
\title{Creates a default watcher configuration file}
\usage{
generate_watcher_config(.mod, watch_opts = list())
}
\arguments{
\item{.mod}{a bbi nonmem model object}

\item{watch_opts}{an optional list to overwrite default values}
}
\value{
none
}
\description{
Creates a default watcher configuration file
}
\examples{
\dontrun{
  generate_watcher_config(.mod)
}
}
\keyword{internal}
