% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-partitions.R
\name{get_slurm_partitions}
\alias{get_slurm_partitions}
\title{Gets the available slurm partitions the user can use.}
\usage{
get_slurm_partitions(cache = TRUE)
}
\arguments{
\item{cache}{optional argument to forgo caching}
}
\value{
vector of available partitions
}
\description{
Gets the available slurm partitions the user can use.
}
\examples{
\dontrun{
get_slurm_partitions
}
}
