% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-partitions.R
\name{lookup_partitions_by_cpu}
\alias{lookup_partitions_by_cpu}
\title{get table of each partition's number of CPUs and memory}
\usage{
lookup_partitions_by_cpu(cache = TRUE)
}
\arguments{
\item{cache}{optional argument to forgo caching}
}
\value{
the processed table of each partition's
number of CPUs and memory
}
\description{
\itemize{
\item gets the raw string output from \code{\link[=run_sinfo]{run_sinfo()}}
\item converts to a data frame
\item reorders and removes \code{*} from default partition with
\code{\link[=process_slurm_partitions]{process_slurm_partitions()}}
}
}
