% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-partitions.R
\name{partition_advice}
\alias{partition_advice}
\title{get partition suggestions}
\usage{
partition_advice(ncpu, partition, avail_cpus_table, cache)
}
\arguments{
\item{ncpu}{number of CPUs requested by user}

\item{partition}{name of partition requested by user}

\item{avail_cpus_table}{table of partitions with respective number of CPUs and memory}

\item{cache}{optional argument to forgo caching}
}
\value{
string with suggestion upon \code{\link[=check_slurm_partitions]{check_slurm_partitions()}} error
}
\description{
In a call to \code{submit_slurm_model()}, if the number of requested CPUs exceeds
the number of CPUs available in the requested partition,
\code{\link[=check_slurm_partitions]{check_slurm_partitions()}} errors. \if{html}{\out{<br />}}
This function follows up with a message providing one or two suggestions for
the partition with the smallest sufficient number of CPUs and least amount
of memory. \if{html}{\out{<br />}}
If there are no partitions with enough CPUs to accommodate the number
requested, this function's return message clarifies this.
}
