% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-partitions.R
\name{process_slurm_partitions}
\alias{process_slurm_partitions}
\title{manipulate partition table for usability}
\usage{
process_slurm_partitions(table)
}
\arguments{
\item{table}{}
}
\value{
the table such that the default partition will be first
and will have the asterisk removed
}
\description{
manipulate partition table for usability
}
