% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-partitions.R
\name{run_sinfo}
\alias{run_sinfo}
\title{get list of each partition's number of CPUs and memory}
\usage{
run_sinfo()
}
\value{
the raw partition-cpu-memory string output from \code{sinfo}
}
\description{
get list of each partition's number of CPUs and memory
}
