% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmtools-package.R
\docType{package}
\name{slurmtools-package}
\alias{slurmtools}
\alias{slurmtools-package}
\title{slurmtools: slurm tooling}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Devin Pastoor \email{devin@a2-ai.com}

Authors:
\itemize{
  \item Jenna Elwing \email{jenna@a2-ai.com}
  \item Matthew Smith \email{matthews@a2-ai.com}
}

}
\keyword{internal}
