% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmtools.R
\docType{package}
\name{slurmtools}
\alias{slurmtools-package}
\alias{slurmtools}
\title{slurmtools: An R package for easily submitting NONMEM jobs to slurm.}
\description{
This package aims to ease the submission and monitoring of NONMEM jobs running
on slurm.
}
\section{submitting jobs}{

\itemize{
\item \code{\link{submit_slurm_job}}: Submits a job to slurm
}
}

\section{monitoring jobs}{

\itemize{
\item \code{\link{get_slurm_jobs}}: Gives a table of submitted jobs to slurm
that shows status and other information given with \code{squeue}
}
}

\section{cancelling jobs}{

\itemize{
\item \code{\link{cancel_slurm_job}}: Cancels the specified job
}
}

\section{slurm partitions}{

\itemize{
\item \code{\link{get_slurm_partitions}}: Gives a vector of available
partitions a user can submit jobs to.
}
}

\author{
\strong{Maintainer}: Devin Pastoor \email{devin@a2-ai.com}

Authors:
\itemize{
  \item Devin Pastoor \email{devin@a2-ai.com}
  \item Jenna Elwing \email{jenna@a2-ai.com}
  \item Matthew Smith \email{matthews@a2-ai.com}
  \item Jacob Dumbleton \email{jacob@a2-ai.com}
}

}
