% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{slurmtools_options_message}
\alias{slurmtools_options_message}
\title{Generates a tidyverse-esque onAttach message}
\usage{
slurmtools_options_message()
}
\value{
a message to display on attach
}
\description{
Generates a tidyverse-esque onAttach message
}
\examples{
slurmtools_options_message()
}
