% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_nonmem_model}
\alias{submit_nonmem_model}
\title{submit a nonmem model to slurm in parallel}
\usage{
submit_nonmem_model(
  .mod,
  partition = get_slurm_partitions(),
  ncpu = 1,
  overwrite = FALSE,
  dry_run = FALSE,
  ...,
  slurm_job_template_path = getOption("slurmtools.slurm_job_template_path"),
  submission_root = getOption("slurmtools.submission_root"),
  bbi_config_path = getOption("slurmtools.bbi_config_path"),
  slurm_template_opts = list()
)
}
\arguments{
\item{.mod}{a path to a model or a bbi nonmem model object}

\item{partition}{name of the partition to submit the model}

\item{ncpu}{number of cpus to run the model against}

\item{overwrite}{whether to overwrite existing model results}

\item{dry_run}{return the command that would have been invoked, without invoking}

\item{...}{arguments to pass to processx::run}

\item{slurm_job_template_path}{path to slurm job template}

\item{submission_root}{directory to track job submission scripts and output}

\item{bbi_config_path}{path to bbi config file}
}
\description{
submit a nonmem model to slurm in parallel
}
