% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_watcher_config.R
\name{validate_toml}
\alias{validate_toml}
\title{Title}
\usage{
validate_toml(toml_list)
}
\arguments{
\item{toml_list}{a list representing the nmm config.toml file}
}
\value{
a potentially trimmed down list containing only actual config values
}
\description{
Title
}
\examples{
\dontrun{
  validate_toml(list(model_number = 1001))
}
}
