% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel-job.R
\name{cancel_slurm_job}
\alias{cancel_slurm_job}
\title{Cancels a running job}
\usage{
cancel_slurm_job(job_id, auto_confirm = FALSE)
}
\arguments{
\item{job_id}{job id to cancel}

\item{auto_confirm}{does not prompt user before cancelling job}
}
\description{
Cancels a running job
}
\examples{
\dontrun{
cancel_job(243)
}
}
