% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel-job.R
\name{cancel_slurm_job}
\alias{cancel_slurm_job}
\title{Cancels a running job}
\usage{
cancel_slurm_job(job_id, confirm = TRUE)
}
\arguments{
\item{job_id}{job id to cancel}

\item{confirm}{requires confirmation to cancel job}
}
\description{
Cancels a running job
}
\examples{
\dontrun{
cancel_job(243)
}
}
