% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmm-config.R
\name{generate_nmm_config}
\alias{generate_nmm_config}
\title{Creates a default watcher configuration file}
\usage{
generate_nmm_config(
  .mod,
  model_number = NULL,
  files_to_track = NULL,
  tmp_dir = NULL,
  watched_dir = file.path("model", "nonmem"),
  output_dir = file.path(watched_dir, "in_progress"),
  poll_duration = NULL,
  level = NULL,
  threads = NULL,
  overwrite = FALSE,
  bbi_config_path = NULL,
  alerter_opts = list()
)
}
\arguments{
\item{.mod}{a bbi nonmem model object}

\item{model_number}{a string of model number e.g. 101a15. Default is pulled from .mod}

\item{files_to_track}{a vector of file extensions. If left blank, nmm will use .ext, .lst. and .grd.}

\item{tmp_dir}{a temporary directory location to run nonmem. If left blank nmm will use /tmp}

\item{watched_dir}{the directory where nonmem will output. Default is here::here()/model/nonmem}

\item{output_dir}{the directory where watcher will place log and any output files default is here::here()/model/nonmem/in_progress}

\item{poll_duration}{the amount of time in seconds between the watchers polling. If left blank, nmm will use 1 second}

\item{level}{What level to log at. Available options are Trace, Debug, Info, Warn, Fatal if left blank, nmm will use Info}

\item{threads}{number of threads if running a parallel job. If left blank nmm will use 1}

\item{overwrite}{bool whether to overwrite existing run.}

\item{bbi_config_path}{the path to bbi.yaml file}

\item{alerter_opts}{a list for setting up an alerter fields needed are alerter, command, message_flag, use_stdout, args (list of additional args to be passed to alerter)}
}
\value{
none
}
\description{
Creates a default watcher configuration file
}
\examples{
\dontrun{
  generate_nmm_config(.mod)
}
}
\keyword{internal}
