% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-partitions.R
\name{get_slurm_partitions}
\alias{get_slurm_partitions}
\title{get list of partition names for the given cluster}
\usage{
get_slurm_partitions(cache = TRUE)
}
\arguments{
\item{cache}{optional argument to forgo caching}
}
\description{
get list of partition names for the given cluster
}
